Create start_here
An easy-to-read overview of purchasing the Osoyoo components, of  constructing the Osoyoo robot, of enhancing the Osoyoo robot,  setting up the correct IDE environment, getting the Osoyoo robot to work, and of interfacing the Osoyoo robot to Olivier Georgeon's project page Python software.
It can be intimidating starting a new project. This overview is intended to give you a gentle approach to the project. More specialized technical details will be posted in other files or see the other documentation on the GitHub repository.